chore: unify actions mentions items
