+++ packages/ui/feature-builder-form-controls/src/lib/interpolating-text-form-control/mentions-list/action-mention-item/action-mention-item.component.html
@@ -3,7 +3,7 @@
     class="ap-flex ap-flex-grow ap-min-h-[27px] ap-items-center">
     <app-generic-step-mention-item [stepIndex]="stepIndex" [stepMention]="stepMention"></app-generic-step-mention-item>
     <div class="ap-flex-grow"></div>
-    <ap-button *ngIf="parentNodeHoverTracker.isHovered && !!(codeStepTest$ | async)" btnColor="primary" btnStyle="basic"
+    <ap-button *ngIf="parentNodeHoverTracker.isHovered && !!(actionTest$ | async)" btnColor="primary" btnStyle="basic"
       type="button" btnSize="small" [buttonIsInsideAutocomplete]="true" [darkLoadingSpinner]="true"
       (click)="$event.stopPropagation();" (buttonClicked)="mentionClicked.emit(stepMention)">Insert
     </ap-button>
@@ -15,12 +15,12 @@
   </div>
 </button>
 
-<ng-container *ngIf="codeStepTest$ | async as result">
+<ng-container *ngIf="actionTest$ | async as result">
   <app-step-mentions-tree *ngIf="result.children && expandCodeCollapse" [stepOutputObjectChildNodes]="result.children"
     [markedNodesToShow]="result.markedNodesToShow" [stepDisplayName]="stepMention.step.displayName"
     (mentionClicked)="emitMention($event)"></app-step-mentions-tree>
 </ng-container>
-<ng-container *ngIf="codeStepTest$ | async as result">
+<ng-container *ngIf="actionTest$ | async as result">
   <button apTrackHover #resultNodeHoverTracker="hoverTrackerDirective" class="!ap-pl-0"
     *ngIf="result.value !==undefined && expandCodeCollapse" mat-menu-item (click)="mentionClicked.emit(stepMention)">
     <span class=" ap-flex-grow ap-transition-all"
@@ -36,11 +36,11 @@
 
 </ng-container>
 
-<ng-container *ngIf="!codeStepTest$">
+<ng-container *ngIf="!actionTest$">
   <ng-container *ngIf="expandCodeCollapse">
     <div class="ap-my-3 ap-px-2">
       <div class="ap-typography-body-1 ap-text-center ap-mb-1" i18n>
-        Test Code First
+        Test Step First
       </div>
       <div class="ap-typography-body-2 ap-text-center ap-mb-3" i18n>
         This step needs to be tested in order to view its data.
@@ -48,7 +48,7 @@
       <div class="ap-text-center">
         <ap-button btnColor="primary" btnStyle="flat" type="button" btnSize="default" [darkLoadingSpinner]="false"
           (buttonClicked)="selectStep();" i18n>
-            Go to step
+          Go to step
         </ap-button>
       </div>
     </div>
