+++ packages/ui/feature-builder-form-controls/src/lib/interpolating-text-form-control/mentions-list/mentions-list.component.html
@@ -24,20 +24,10 @@
 
   <ng-container *ngFor="let stepMention of stepsMentions$|async" #stepsListTemplate>
     <ng-container [ngSwitch]="stepMention.step.type">
-      <ng-container *ngSwitchCase="ActionType.CODE">
-        <!---TODO FIX -->
-        <app-code-step-mention-item [stepIndex]="stepMention.step.indexInDfsTraversal"
-          (mentionClicked)="mentionClicked($event)" [stepMention]="stepMention"></app-code-step-mention-item>
-      </ng-container>
-      <ng-container *ngSwitchCase="ActionType.PIECE">
-        <!---TODO FIX -->
-        <app-piece-step-mention-item [stepIndex]="stepMention.step.indexInDfsTraversal"
-          (mentionClicked)="mentionClicked($event)" [stepMention]="stepMention"></app-piece-step-mention-item>
-      </ng-container>
-      <ng-container *ngSwitchCase="ActionType.LOOP_ON_ITEMS">
-        <!---TODO FIX -->
-        <app-loop-step-mention-item [stepIndex]="stepMention.step.indexInDfsTraversal"
-          (mentionClicked)="mentionClicked($event)" [stepMention]="stepMention"></app-loop-step-mention-item>
+      <ng-container
+        *ngIf="stepMention.step.type === ActionType.CODE || stepMention.step.type === ActionType.LOOP_ON_ITEMS ||  stepMention.step.type === ActionType.PIECE">
+        <app-action-mention-item [stepIndex]="stepMention.step.indexInDfsTraversal"
+          (mentionClicked)="mentionClicked($event)" [stepMention]="stepMention"></app-action-mention-item>
       </ng-container>
 
       <ng-container *ngSwitchCase="TriggerType.PIECE">
