+++ packages/ui/feature-builder-test-steps/src/lib/test-action/test-action.component.ts
@@ -221,8 +221,8 @@ export class TestActionComponent extends TestStepCoreComponent {
                   'index' in testResult
                 ) {
                   const item = {
-                    item: testResult['item'],
                     index: testResult['index'],
+                    item: testResult['item'],
                   };
                   const clone: LoopOnItemsAction = {
                     ...step,
